% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci.R
\name{naive_fibonacci}
\alias{naive_fibonacci}
\title{A function to compute the nth Fibonacci number}
\usage{
naive_fibonacci(n, accumulate = FALSE)
}
\arguments{
\item{n}{An integer representing the desired element of the series to return.}
}
\description{
Computes the nth Fibonacci recursively. This becomes computationally
intensive for larger values of n.
}
